+++ include/_Release.h
@@ -1,1 +1,1 @@
-#define Release "3.4.1"
+#define Release "3.4.2"